Refactor cryptography class to move all state agnostic methods to encryptable module
