+++ lib/cryptography.rb
@@ -2,8 +2,10 @@ require 'date'
 require './lib/alphabet.rb'
 require './lib/key'
 require './lib/offset.rb'
+require './lib/encryptable.rb'
 
 class Cryptography
+  include Encryptable
 
   attr_reader :alphabet
 
@@ -25,9 +27,7 @@ class Cryptography
     }
   end
 
-  def convert_message(message)
-    message.downcase.split(//)
-  end
+
 
   def shift_characters(message_characters)
     counter = 0
@@ -55,13 +55,6 @@ class Cryptography
     @alphabet.alpha_by_number[number]
   end
 
-  def find_shift(counter)
-    return :A if counter % 4 == 1
-    return :B if counter % 4 == 2
-    return :C if counter % 4 == 3
-    :D
-  end
-
   def calculate_shifts
     keys = Key.key_values
     offsets = ShiftOffset.offset_values
@@ -70,26 +63,6 @@ class Cryptography
       shift
     end
     @shifts
-  end
-
-  def split_characters(message)
-    message.split()
-  end
-
-  def date_conversion(date)
-    date.strftime("%m%d%Y")
-  end
-
-  def random_key
-    rand(999999)
-  end
-
-  def pad_key(rando_key)
-    rando_key = rando_key.to_s
-    while rando_key.length < 5
-      rando_key.prepend("0")
-    end
-    rando_key
   end
 
 end
